Refactor taskflow and resource optimization in MLIR tests
- Updated taskflow definitions in `parallel-nested.mlir` to include `cgra_shape` attribute for better resource management.
- Modified `simple_resnet_tosa.mlir` to replace `tile_shape` with `cgra_shape` in taskflow tasks, enhancing clarity and consistency across resource-aware optimizations.
- Removed `resource-heavy.mlir` as it was deemed unnecessary, streamlining the test suite.
