+++ include/TaskflowDialect/TaskflowPasses.h
@@ -21,6 +21,7 @@ void registerTosaToAffineConversionPassPipeline();
 // Passes defined in TaskflowPasses.td
 #define GEN_PASS_DECL
 #include "TaskflowDialect/TaskflowPasses.h.inc"
+std::unique_ptr<mlir::Pass> createAllocateCgraToTaskPass();
 std::unique_ptr<mlir::Pass> createConstructHyperblockFromTaskPass();
 std::unique_ptr<mlir::Pass> createClassifyCountersPass();
 std::unique_ptr<mlir::Pass> createAllocateCgraToTaskPass();
