+++ lib/TaskflowDialect/CMakeLists.txt
@@ -15,3 +15,4 @@ add_mlir_dialect_library(MLIRTaskflow
 
 add_subdirectory(Allocation)
 add_subdirectory(Transforms)
+add_subdirectory(Transforms/Optimizations)
