+++ lib/TaskflowDialect/Transforms/CMakeLists.txt
@@ -1,6 +1,7 @@
 get_property(dialect_libs GLOBAL PROPERTY MLIR_DIALECT_LIBS)
 
 add_mlir_library(MLIRTaskflowTransforms
+    AllocateCgraToTaskPass.cpp
     ConstructHyperblockFromTaskPass.cpp
     ClassifyCountersPass.cpp
     AllocateCgraToTaskPass.cpp
